requirements.txt
python-telegram-bot
python-dotenv
gspread
oauth2client
openai
json
requests
langchain
langchain_community
re 
tiktoken

numpy >= 1.12.1
scipy >= 1.0.1
torch >= 2.0.0
accelerate == 0.18.0
bitsandbytes == 0.37.2
git+https://github.com/huggingface/peft.git@8e53e16005
transformers == 4.29.1
pillow==9.5.0
llama-cpp-python
datasets
zstandard
jsonlines
wandb
openai
sentencepiece
fire
datasketch==1.5.9
